apps/api/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.28.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # 必须与 passlib 兼容，4.1+ 或 5.0+ 可能不兼容

# Database
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pymongo==4.6.2
motor==3.3.2
redis==5.0.3

# AI/ML - Core
openai==1.14.0
httpx==0.27.0

# Task Queue
celery==5.3.6

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Data Processing
pandas==2.2.1
numpy==1.26.4

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
tenacity==8.2.3

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0

# Auth & OAuth
python-jose[cryptography]==3.3.0
httpx==0.27.0

